Move RPCMessageType to enums
