+++ freqtrade/rpc/__init__.py
@@ -1,3 +1,3 @@
 # flake8: noqa: F401
-from .rpc import RPC, RPCException, RPCHandler, RPCMessageType
+from .rpc import RPC, RPCException, RPCHandler
 from .rpc_manager import RPCManager
