+++ tests/rpc/test_rpc_webhook.py
@@ -5,8 +5,8 @@ from unittest.mock import MagicMock
 import pytest
 from requests import RequestException
 
-from freqtrade.enums import SellType
-from freqtrade.rpc import RPC, RPCMessageType
+from freqtrade.enums import RPCMessageType, SellType
+from freqtrade.rpc import RPC
 from freqtrade.rpc.webhook import Webhook
 from tests.conftest import get_patched_freqtradebot, log_has
